requirements.txt
acres==0.2.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
argcomplete==1.10.3
attrs==24.2.0
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
compressed-rtf==1.0.6
configobj==5.0.9
configparser==7.1.0
defusedxml==0.7.1
distro==1.9.0
django-environ==0.11.2
docx2txt==0.8
ebcdic==1.1.1
environ==1.0
etelemetry==0.3.1
extract-msg==0.28.7
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
importlib_resources==6.4.5
isodate==0.6.1
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
language_tool_python==2.8.1
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.18.3
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==2.2.0
odfpy==1.4.1
olefile==0.47
openai==1.57.4
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20191110
pillow==11.0.0
protobuf==5.29.1
prov==2.0.1
puremagic==1.28
pyarrow==18.1.0
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.3
Pygments==2.18.0
PyMuPDF==1.25.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.14.1
shellingham==1.5.4
simplejson==3.19.3
six==1.12.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
streamlit==1.41.0
tenacity==9.0.0
textract==1.6.5
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==6.4.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
watchdog==6.0.0
xlrd==1.2.0
XlsxWriter==3.2.0
